basic_agent/region_agent/requirements.txt
google-adk
pandas
numpy